Remove SuperSpacePixel and make SpacePixelGroup non-generic
As the only implementation that remains after the last commit is SpacePixelGroup<ShapeMap> which was equivalent to SpacePixelFile it has been renamed as such and made non-generic to simplify inheritance
